Face: Add Send+Sync bound to closure
HarfBuzz Objects are thread-safe. This means that the closure provided
in the from_table_func method of Face should be able to be called from
any thread.

At the moment I am not sure if the Sync bound is really needed. However
to be safe this commit adds a Sync bound to the closure which may be
removed later if deemed safe.