Add Sync bound to Font::set_font_funcs
There is currently nothing preventing you from calling the font funcs
asynchronously as `Font` implements both `Send` and `Sync`. It remains
to be seen if this bound needs to be relaxed by some means in the
future.

This prevents that implementors of `FontFuncs` contain `Cell`s or
`RefCells` when used with a `Font`. When interior mutability is needed
(e.g. for caching) you need to use `Mutex`es or similar constructs.