Change return type of Font::parent() and check for null pointer
This is a required breaking change to make the API exposed by rust safe.

Before this commit there was a possibility to dereference a null pointer
when trying to get the parent of the empty font (which is the default
parent font of a new harfbuzz font).

E.g. the following code snippet could cause undefined behaviour:

let font = Font::new(face);
let parent = font.parent(); //< this is ok
let grand_parent = parent.parent(); //< this is UB

Therefore we have to check the output of hb_font_get_parent and change
the return type of Font::parent() to Option<&Font> (which is a breaking
change).