Fix lifetime of slice returned by Blob
The `get_data` method on `Blob` until this commit returned a slice with
the same lifetime as the Blob itself. However this is wrong and can lead
to undefined behaviour in safe code.

The problem was that the explicit specification of lifetime on the
returned slice meant that the Blob was not borrowed whilst the returned
slice was active. Then if the Blob were dropped you could still access
the slice's data which is obviously undefined behaviour.

This mistake is most obvious when you consider the 'static lifetime.
Then `get_data()` would return a `static slice which is flawed because
a static slice would need to live forever which Blob can't ensure.

The fix to `get_data` also introduces changes in the API of `Face`.
Instead of returning slices the relevant function now directly return
`Blob`s. This way the caller is actually in charge of dereferencing the
`Blob` and can extend its lifetime as long as needed.