Implement glyph buffer serialization
HarfBuzz provides a function to serialize the contents of a glyph
buffer.

This capability is now exposed as a rustic API introducing a new type
`BufferSerializer`. A value of this type can be obtained through the
new `GlyphBuffer::serializer` function. A `BufferSerializer` imlements
`std::io::Read` to allow users to get the serialized representation of
the glyph buffer.

Deserialization is future work.